Discord_Bot/requirements.txt
openpyxl
discord.py
python-dotenv
matplotlib
